requirements.txt
cartopy>=0.24.1
cftime>=1.6.4.post1
h5netcdf>=1.6.1
netCDF4>=1.7.0
matplotlib>=3
numpy>=1.25.0
pandas>=2.2.3
tqdm>=4.67.1
xarray>=2025.3.1
pyyaml>=6.0
scipy>=1.15.3
